tasks/speech/text2speech/fastspeech2_HS/pyproject.toml
[project]
name = "fastspeech2-hs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "espnet>=202511",
    "fastapi>=0.125.0",
    "huggingface-hub>=1.2.3",
    "indic-num2words>=1.3.2",
    "indic-unified-parser>=1.0.6",
    "librosa>=0.11.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "noisereduce>=3.0.3",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "psutil>=7.1.3",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "uvicorn>=0.40.0",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchaudio = [
  { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
